feat: get rid of tailwindcss cdn, generate css locally
NOTE: In this commit we made a lot of changes, some possibly creating
bugs. Main problem was the fact that tailwind cli cannot generate
dynamically created classes (like `bg-[%s]` in templ), it needs to 'see'
them in full in the source files. To solve this, we had to create css
variables for all the dynamic classes we used, and then use and change
those variables accordingly.

NOTE2: Follow the installation instructions from tailwindcss website to
get the cli working locally. After that, run `npx @tailwindcss/cli -i assets/css/input.css -o assets/css/style.css --watch` to generate the css file locally. Make sure to keep the process running in the background while developing, so that the css file is always up to date.